Refactor Instrumented annotation into the agent.
