+++ quasar-core/src/main/java/co/paralleluniverse/fibers/instrument/Classes.java
@@ -21,6 +21,7 @@ import java.util.HashSet;
 import java.util.EnumMap;
 import java.util.Collections;
 
+import co.paralleluniverse.fibers.suspend.Instrumented;
 import org.objectweb.asm.Type;
 import org.objectweb.asm.tree.MethodInsnNode;
 
@@ -57,7 +58,7 @@ final class Classes {
     static final String LAMBDA_METHOD_PREFIX            = "lambda$";
 
     static final String DONT_INSTRUMENT_DESC = Type.getDescriptor(DontInstrument.class);
-    static final String INSTRUMENTED_DESC    = "Lco/paralleluniverse/fibers/Instrumented;";
+    static final String INSTRUMENTED_DESC    = Type.getDescriptor(Instrumented.class);
 
     // CORE-21 : Provide getter and setter for annotation types.
     static class AnnotationDescriptors {
@@ -67,7 +68,7 @@ final class Classes {
         };
 
         // Keep as non volatile for now as should only be modified by the agent on initialisation.
-        private EnumMap<ID, Set<String>> descIds = new EnumMap<>(ID.class);
+        private final EnumMap<ID, Set<String>> descIds = new EnumMap<>(ID.class);
 
         AnnotationDescriptors() {
             // We use string literals rather than Type.getDescriptor as we do not want
