+++ quasar-core/src/main/java/co/paralleluniverse/fibers/suspend/Instrumented.java
@@ -39,7 +39,7 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
-package co.paralleluniverse.fibers;
+package co.paralleluniverse.fibers.suspend;
 
 import co.paralleluniverse.fibers.instrument.Constants;
