+++ quasar-core/src/test/java/co/paralleluniverse/fibers/instrument/BlockingTest.java
@@ -28,7 +28,7 @@
  */
 package co.paralleluniverse.fibers.instrument;
 
-import co.paralleluniverse.fibers.Instrumented;
+import co.paralleluniverse.fibers.suspend.Instrumented;
 import co.paralleluniverse.fibers.suspend.SuspendExecution;
 import java.io.IOException;
 import java.io.InputStream;
