+++ quasar-core/src/test/java/co/paralleluniverse/fibers/instrument/InstrumentationOptimizerTest.java
@@ -14,7 +14,7 @@
 package co.paralleluniverse.fibers.instrument;
 
 import co.paralleluniverse.fibers.Fiber;
-import co.paralleluniverse.fibers.Instrumented;
+import co.paralleluniverse.fibers.suspend.Instrumented;
 import co.paralleluniverse.fibers.Suspendable;
 import co.paralleluniverse.fibers.suspend.SuspendExecution;
 import co.paralleluniverse.strands.SuspendableRunnable;
